requirements.txt
numpy==2.1.1
pandas==2.2.3
protobuf==5.28.2
sentence_transformers==3.1.1
streamlit==1.38.0
streamlit_ext==0.1.10
torch==2.4.1
google-cloud-storage
